+++ data-raw/random_mctq.R
@@ -70,23 +70,47 @@ std_mctq_par <- function() {
 
         ms_mean <- i$ms_mean
         ms_sd <- i$ms_sd
-        ms_min <- sum_time(ms_mean, - (3 * ms_sd), cycle = lubridate::ddays())
-        ms_max <- sum_time(ms_mean, + (3 * ms_sd), cycle = lubridate::ddays())
+        ms_min <- sum_time(ms_mean, - (3 * ms_sd),
+                           cycle = lubridate::ddays()) %>%
+            as.numeric() %>%
+            hms::hms()
+        ms_max <- sum_time(ms_mean, + (3 * ms_sd),
+                           cycle = lubridate::ddays()) %>%
+            as.numeric() %>%
+            hms::hms()
 
         sd_mean <- i$sd_mean
         sd_sd <- i$sd_sd
-        sd_min <- sum_time(sd_mean, - (3 * sd_sd), cycle = lubridate::ddays())
-        sd_max <- sum_time(sd_mean, + (3 * sd_sd), cycle = lubridate::ddays())
+        sd_min <- sum_time(sd_mean, - (3 * sd_sd),
+                           cycle = lubridate::ddays()) %>%
+            as.numeric() %>%
+            hms::hms()
+        sd_max <- sum_time(sd_mean, + (3 * sd_sd),
+                           cycle = lubridate::ddays()) %>%
+            as.numeric() %>%
+            hms::hms()
 
         cli::cli_h3(paste0(
             "Local time of sleep onset (SO_", i$suffix, ")"
             ))
         cli::cat_line()
 
-        mean <- sum_time(ms_mean, - (sd_mean / 2), cycle = lubridate::ddays())
-        sd <- sum_time((ms_sd + sd_sd) / 2, cycle = lubridate::ddays())
-        min <- sum_time(ms_min, - (sd_mean / 2), cycle = lubridate::ddays())
-        max <- sum_time(ms_max, - (sd_mean / 2), cycle = lubridate::ddays())
+        mean <- sum_time(ms_mean, - (sd_mean / 2),
+                         cycle = lubridate::ddays()) %>%
+            as.numeric() %>%
+            hms::hms()
+        sd <- sum_time((ms_sd + sd_sd) / 2,
+                       cycle = lubridate::ddays()) %>%
+            as.numeric() %>%
+            hms::hms()
+        min <- sum_time(ms_min, - (sd_mean / 2),
+                        cycle = lubridate::ddays()) %>%
+            as.numeric() %>%
+            hms::hms()
+        max <- sum_time(ms_max, - (sd_mean / 2),
+                        cycle = lubridate::ddays()) %>%
+            as.numeric() %>%
+            hms::hms()
         cat_(min, max, mean, sd)
 
         cli::cli_h3(paste0(
@@ -94,10 +118,26 @@ std_mctq_par <- function() {
         ))
         cli::cat_line()
 
-        mean <- sum_time(ms_mean, + (sd_mean / 2), cycle = lubridate::ddays())
-        sd <- sum_time((ms_sd + sd_sd) / 2, cycle = lubridate::ddays())
-        min <- sum_time(ms_min, + (sd_mean / 2), cycle = lubridate::ddays())
-        max <- sum_time(ms_max, + (sd_mean / 2), cycle = lubridate::ddays())
+        mean <- sum_time(ms_mean, + (sd_mean / 2),
+                         cycle = lubridate::ddays()) %>%
+            as.numeric() %>%
+            hms::hms()
+
+        sd <- sum_time((ms_sd + sd_sd) / 2,
+                       cycle = lubridate::ddays()) %>%
+            as.numeric() %>%
+            hms::hms()
+
+        min <- sum_time(ms_min, + (sd_mean / 2),
+                        cycle = lubridate::ddays()) %>%
+            as.numeric() %>%
+            hms::hms()
+
+        max <- sum_time(ms_max, + (sd_mean / 2),
+                        cycle = lubridate::ddays()) %>%
+            as.numeric() %>%
+            hms::hms()
+
         cat_(min, max, mean, sd)
 
         cli::cli_h3(paste0(
@@ -264,10 +304,17 @@ shift_mctq_par <- function() {
 
         sprep_mean <- i$sprep_mean
         sprep_sd <- i$sprep_sd
+
         sprep_min <- sum_time(sprep_mean, - (3 * sprep_sd),
-                              cycle = lubridate::ddays())
+                              cycle = lubridate::ddays()) %>%
+            as.numeric() %>%
+            hms::hms()
+
         sprep_max <- sum_time(sprep_mean, + (3 * sprep_sd),
-                              cycle = lubridate::ddays())
+                              cycle = lubridate::ddays()) %>%
+            as.numeric() %>%
+            hms::hms()
+
         cat_(sprep_min, sprep_max, sprep_mean, sprep_sd)
 
         cli::cli_h3(paste0(
@@ -277,9 +324,17 @@ shift_mctq_par <- function() {
 
         slat_mean <- i$slat_mean
         slat_sd <- i$slat_sd
-        slat_min <- sum_time(slat_mean, - (3 * slat_sd))
+
+        slat_min <- sum_time(slat_mean, - (3 * slat_sd)) %>%
+            as.numeric() %>%
+            hms::hms()
+
         if (slat_min <= 0) slat_min <- hms::as_hms(0)
-        slat_max <- sum_time(slat_mean, + (3 * slat_sd))
+
+        slat_max <- sum_time(slat_mean, + (3 * slat_sd)) %>%
+            as.numeric() %>%
+            hms::hms()
+
         cat_(slat_min, slat_max, slat_mean, slat_sd)
 
         cli::cli_h3(paste0(
@@ -287,10 +342,22 @@ shift_mctq_par <- function() {
         ))
         cli::cat_line()
 
-        mean <- sum_time(sprep_mean, slat_mean, cycle = lubridate::ddays())
-        sd <- sum_time(sprep_sd + slat_sd)
-        min <- sum_time(sprep_min, slat_min, cycle = lubridate::ddays())
-        max <- sum_time(sprep_max, slat_max, cycle = lubridate::ddays())
+        mean <- sum_time(sprep_mean, slat_mean, cycle = lubridate::ddays()) %>%
+            as.numeric() %>%
+            hms::hms()
+
+        sd <- sum_time(sprep_sd + slat_sd) %>%
+            as.numeric() %>%
+            hms::hms()
+
+        min <- sum_time(sprep_min, slat_min, cycle = lubridate::ddays()) %>%
+            as.numeric() %>%
+            hms::hms()
+
+        max <- sum_time(sprep_max, slat_max, cycle = lubridate::ddays()) %>%
+            as.numeric() %>%
+            hms::hms()
+
         cat_(min, max, mean, sd)
 
         cli::cli_h3(paste0(
@@ -309,9 +376,17 @@ shift_mctq_par <- function() {
 
         mean <- i$tgu_mean
         sd <- i$tgu_sd
-        min <- sum_time(mean, - (3 * sd))
+
+        min <- sum_time(mean, - (3 * sd)) %>%
+            as.numeric() %>%
+            hms::hms()
+
         if (min <= 0) min <- hms::as_hms(0)
-        max <- sum_time(mean, + (3 * sd))
+
+        max <- sum_time(mean, + (3 * sd)) %>%
+            as.numeric() %>%
+            hms::hms()
+
         cat_(min, max, mean, sd)
 
         cli::cli_h3(paste0(
@@ -625,8 +700,13 @@ cat_ <- function(min, max, mean, sd) {
 }
 
 min_max <- function(mean, sd) {
-    min <- sum_time(mean, - (3 * sd), cycle = lubridate::ddays())
-    max <- sum_time(mean, + (3 * sd), cycle = lubridate::ddays())
+    min <- sum_time(mean, - (3 * sd), cycle = lubridate::ddays()) %>%
+        as.numeric() %>%
+        hms::hms()
+
+    max <- sum_time(mean, + (3 * sd), cycle = lubridate::ddays()) %>%
+        as.numeric() %>%
+        hms::hms()
 
     cat_(min, max, mean, sd)
 }